pyproject.toml
[project]
name = "crossy-road"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "neat-python>=1.1.0",
    "pygame>=2.6.1",
    "streamlit>=1.53.1",
]
